Merged all patches from DEBIAN except 01build_all_python_versions.dpatch
This resulted in a name change of the shared library, the header file names and the library SONAME tag from dnet to dumbnet due to a conflict with the much older DECnet library on DEBIAN.

So software built to look fro dnet will not run with these change.
This is a pretty big deal. I'm not sure if I really want that.
Until I do, this commit will be on its own branch.